+++ src/js/components/TabPanesComponent.jsx
@@ -1,3 +1,4 @@
+var Link = require("react-router").Link;
 var React = require("react/addons");
 
 var AppListFilterComponent = require("../components/AppListFilterComponent");
@@ -8,8 +9,7 @@ var DeploymentsListComponent =
 var SidebarComponent = require("../components/SidebarComponent");
 var TabPaneComponent = require("../components/TabPaneComponent");
 var TogglableTabsComponent = require("../components/TogglableTabsComponent");
-
-var AppListViewTypes = require("../constants/AppListViewTypes");
+var Util = require("../helpers/Util");
 var tabs = require("../constants/tabs");
 
 var TabPanesComponent = React.createClass({
@@ -34,6 +34,34 @@ var TabPanesComponent = React.createClass({
     this.updateCurrentGroup();
   },
 
+  getContextualBar: function () {
+    var state = this.state;
+
+    if (state.filters.filterText == null ||
+        Util.isEmptyString(state.filters.filterText)) {
+      return <BreadcrumbComponent groupId={state.currentGroup} />;
+    }
+
+    let router = this.context.router;
+    let currentPathname = router.getCurrentPathname();
+    let query = Object.assign({}, router.getCurrentQuery());
+    let params = router.getCurrentParams();
+
+    delete query.filterText;
+
+    return (
+        <p className="breadcrumb">
+          <span>{`Search results for "${state.filters.filterText}"`}</span>
+          <Link className="clear"
+              to={currentPathname}
+              query={query}
+              params={params}>
+            Clear search
+          </Link>
+        </p>
+    );
+  },
+
   updateCurrentGroup: function () {
     var {groupId} = this.context.router.getCurrentParams();
     if (groupId == null) {
@@ -50,9 +78,10 @@ var TabPanesComponent = React.createClass({
   },
 
   updateFilters: function (filters) {
-    var updatedFilters = Object.assign({}, this.state.filters, filters);
-    this.setState({
-      filters: updatedFilters
+    this.setState(function (prevState) {
+      return {
+        filters: Object.assign({}, prevState.filters, filters)
+      };
     });
   },
 
@@ -71,11 +100,6 @@ var TabPanesComponent = React.createClass({
   render: function () {
     var state = this.state;
 
-    var appListProps = Object.assign({
-      currentGroup: state.currentGroup,
-      viewType: AppListViewTypes.LIST
-    }, state.filters);
-
     return (
       <TogglableTabsComponent activeTabId={this.getTabId()}
           className="container-fluid content">
@@ -85,12 +109,13 @@ var TabPanesComponent = React.createClass({
               onChange={this.updateFilters} />
             <main>
               <div className="contextual-bar">
-                <BreadcrumbComponent groupId={state.currentGroup} />
+                {this.getContextualBar()}
                 <div className="app-list-controls">
                   <AppListFilterComponent onChange={this.updateFilters} />
                 </div>
               </div>
-              <AppListComponent {...appListProps} />
+              <AppListComponent currentGroup={state.currentGroup}
+                filters={state.filters} />
             </main>
           </div>
         </TabPaneComponent>