+++ src/js/stores/AppsStore.js
@@ -320,6 +320,9 @@ AppDispatcher.register(function (action) {
         action.taskIds
       );
       break;
+    case AppsEvents.UPDATE_APPS_STATUSES_COUNT:
+      AppsStore.emit(AppsEvents.UPDATE_APPS_STATUSES_COUNT, action.data);
+      break;
   }
 });