+++ CHANGELOG.md
@@ -17,6 +17,9 @@
 - \#2662 - Change Header When Applying Filter in App Collection View
 - \#2663 - Adapt filter count, reflect current result set
 
+### Fixed
+- \#2691 - Filters don't updated correctly on reload
+
 ## 0.13.4 - 2015-11-18
 ### Fixed
 - \#2626 - Status icons are rendered blotted